Stub map tiles in Playwright tests
