+++ frontend/e2e/fixtures.js
@@ -1,0 +1,44 @@
+const { expect, test: base } = require("@playwright/test");
+
+const transparentPng = Buffer.from(
+	"iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8z8BQDwAFgwJ/lz7eSgAAAABJRU5ErkJggg==",
+	"base64",
+);
+
+const test = base.extend({
+	page: async ({ page }, use) => {
+		const consoleErrors = [];
+		const failedRequests = [];
+
+		await page.route("https://tile.openstreetmap.org/**", async (route) => {
+			await route.fulfill({
+				status: 200,
+				contentType: "image/png",
+				body: transparentPng,
+			});
+		});
+
+		page.on("console", (message) => {
+			if (message.type() === "error") {
+				consoleErrors.push(message.text());
+			}
+		});
+
+		page.on("requestfailed", (request) => {
+			const failure = request.failure();
+			failedRequests.push(
+				`${request.method()} ${request.url()} ${failure?.errorText ?? "failed"}`,
+			);
+		});
+
+		await use(page);
+
+		expect(consoleErrors).toEqual([]);
+		expect(failedRequests).toEqual([]);
+	},
+});
+
+module.exports = {
+	expect,
+	test,
+};
